+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `CTRL + Mouse wheel` or `CTRL + +/-` to zoom in/out on console, editor, help.
 - toolbar for figure with print, zoom in, zoom out, rotation, pan, restore axes.
 - `MenuBar`, `ToolBar` figure properties.
+- `View` property for Axes.
 - Window menu on graphic window, list all others available windows.
 - `feature` builtin (undocument features, debug, tests, ...) content can change with next releases.
 
@@ -24,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Clicking on an axis automatically sets it as the current axes object.
 - Clicking on an figure automatically sets it as the current figure object.
 - `saveas` exports the figure as a PDF page with centered alignment.
+- view function returns azimuth and elevation values.
+- camera view reworked.
+- default figure size updated.
 
 ### Fixed